refactor: refactor pet, grid(now called petgrid) and utils modules
